Build 2.0.0.41
---"Battle Engineer: Thunder Shock Weapon" DCs are now correctly Mod ability, not full ability (Reported by Spook)
---"Battle Engineer: Shatter Defenses" DCs are now correctly Mod ability, not full ability (Reported by Spook)
---"Renegade Mastermaker: Mighty Slam" DC also fixed
---"Melee Alacrity" should now correctly apply to melee weapons (Reported by Refutor)
---Ravager cores 4/5/6 now correctly add +2 Intimidate (descriptions also updated) (Reported by Kalibano)
---Feat "Epic Reputation" is now correctly +5 for listed skills (Reported by Kalibano)
---Augment "Crystallized Drop of Tea" will now correctly reduce the cost of Empower Spell metamagic (Reported by Viperswhip)
---The "Bloodrage Chrism" effects will now have the correct bonus types (Reported by Kalibano)
---Feat "Magical Training" is now correctly limited for Paladin, rangers and Bards required min levels (Reported by MuazAlhaidar)
---Set "Epic Vulkoor's Might" and "Legendary Vulkoor's Might" now have all their set bonus entries (Reported by Mummykiller)
---Feat "Arcane Actions" imbue dice bonus is now correctly a flat +2 dice bonus, not a +2% bonus (Reported by Mummykiller)
---Set bonus augment "Truthful Blow" is now correctly a 15% bonus, not a 30% bonus (Reported by Mummykiller)
---"Shiradi Champion: Feywild Attunement" should now correctly set BAB to character level (or 25, whichever is lowest) (Reported by Mummykiller)
---"MRR Cap" for War/Bladeforged characters will now be applied correctly based on your body feat (Reported by Mummykiller)
---"Dodge Cap" for War/Bladeforged characters will now be applied correctly based on your body feat (Reported by Mummykiller)
---"Composite Plating" armor bonus for some docents should not accidentally hit MAX_INT due to signed/unsigned issues (Reported by Mummykiller)
---Verify effects issue fixed so spell effects which have MAX_BUILDER_LEVEL now they expect 41 elements not 40.
---Wiki download redone. Many missing "Golden Age" weapons and "Eon, the Poisoned Reign"" weapon added.
---Item "The Winter Solstice" fixed (Reported by Mummykiller)
---All "Legendary Ring of Mystic <Ability>" items now have their Quality Deception +3 buff (Reported by Mummykiller)
---Several bugs in the Wiki item processing of "Deception" effects was fixed. All items reprocessed.
---A middle click on the feats select control will now allow an alternate feat selection if you have a regular feat already selected.
------Alternate feats now appear on the Forum export in a different colour (when BB code enabled)
---Item "Legendary Powder-Packed Barrel" now has its missing set bonus (Reported by Bizmarkie and Bjond)
---Augment "Topaz of Anthem" no longer erroneously grants a stack of set "Anthem Melody" (Reported by Bjond)
---Archetype past lives from V1 files will now import correctly (Reported by Bjond)
---Buff "Enhanced Ghostly" will now grant its "Incorporeal" bonus of 15% (Reported by Bjond)
---"Epic Quiver of Alacrity" now has the correct alacrity buff (Reported by Bjond)
---"Dragon Lord: Untouchable Soul" displacement value will now show in breakdowns (bad effect Concealment removed) (Reported by Bjond)
---Item "Dinosaur Bone Buckler" now only has one instance of its set bonus (Reported by Bjond)
---Item "Legendary Skullduggery Cap" now has its missing set bonus (Reported by Bjond)
---"Unyielding Sentinel: Hardened" will now correctly add its 10% energy absorbance to mantle at rank 3 (Reported by Bjond)
---Enhancements "Wood Elf: Nature I/II" no longer erroneously grant an extra +1 dexterity (Reported by Bjond)
---"Implement in your hands" will now only award "Enhancement bonus" Implement bonus rather than *3 as per https://ddowiki.com/page/Implement_bonus (Reported by Bjond)
---The warning for a required feat swap at level 1 due to stat tomes not applying has returned
------Also appears in forum export
---Layout issue in skills export fixed
---Item "Legendary Goggles of the Berserker" "Enhanced Bloodrage" effects will now have the correct bonus types (Reported by christhemiss)
---Buff "Lesser Displacement" is now correctly an Enhancement bonus (Reported by Ozmandias42)
---Spell "Spell Resistance, Mass" is now correctly an Enhancement bonus (Reported by Ozmandias42)
---Bonus %Shield AC should now work again (Reported by Ozmandias42)
---Various other spell effects which gave an "Enchantment" bonus converted to "Enhancement"
---"Shadow Shrouding - Concealment/Incorporeality" will no longer erroneously stack with items (Reported by Kalibano)
---Spell effect carry over from trained spells on class change will no longer occur (Reported by Kalibano)
---"Ever Watchful" feat PRR effect now correctly scales with the number of "Religious Lore" feats you have (Reported by Kalibano)

+++ DDOBuilder/Item.h
@@ -76,9 +76,9 @@ class Item :
                 DL_OPTIONAL_OBJECT(_, BaseDice, DamageDice) \
                 DL_OPTIONAL_SIMPLE(_, size_t, CriticalMultiplier, 0) \
                 DL_OPTIONAL_SIMPLE(_, size_t, CriticalThreatRange, 0) \
-                DL_OPTIONAL_SIMPLE(_, size_t, ArmorBonus, 0) \
-                DL_OPTIONAL_SIMPLE(_, size_t, MithralBody, 0) \
-                DL_OPTIONAL_SIMPLE(_, size_t, AdamantineBody, 0) \
+                DL_OPTIONAL_SIMPLE(_, int, ArmorBonus, 0) \
+                DL_OPTIONAL_SIMPLE(_, int, MithralBody, 0) \
+                DL_OPTIONAL_SIMPLE(_, int, AdamantineBody, 0) \
                 DL_OPTIONAL_SIMPLE(_, size_t, MaximumDexterityBonus, 0) \
                 DL_OPTIONAL_SIMPLE(_, int, ArmorCheckPenalty, 0) \
                 DL_OPTIONAL_SIMPLE(_, int, ArcaneSpellFailure, 0) \
